+++ paillmtrace-20240311/ChangeLog.md
@@ -1,3 +1,8 @@
+2025-08-02 Version: 1.1.2
+- Update API ListOnlineEvalTasks: add request parameters AppName.
+- Update API ListOnlineEvalTasks: add request parameters Status.
+
+
 2025-08-01 Version: 1.1.1
 - Update API ListOnlineEvalTasks: add request parameters SortBy.
 - Update API ListOnlineEvalTasks: add request parameters SortOrder.